quinn/requirements.txt
Flask==3.0.0
PyYAML==6.0.1
Authlib==1.3.0
python-dotenv==1.0.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
